requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
azure-functions==1.18.0
azure-identity==1.15.0
requests==2.31.0
python-dotenv==1.0.0
